+++ src/app/modules/shelf/shelf-management/shelf-management.component.ts
@@ -1,87 +1,88 @@
 import { Component, OnInit } from '@angular/core';
-import { Shelf } from '../../../shared/dto/shelf';
 import { ActivatedRoute, Router } from '@angular/router';
-import { ProductService } from '../../../shared/service/product.service';
 import { ToastrService } from 'ngx-toastr';
 import { ShelfService } from '../../../shared/service/shelf.service';
-import { HttpClient } from '@angular/common/http';
 import { AdminShelf } from '../../../shared/dto/admin-shelf';
 import { AdminProduct } from '../../../shared/dto/admin-product';
+import { FormBuilder } from '@angular/forms';
 
 @Component({
   selector: 'app-shelf-management',
   templateUrl: './shelf-management.component.html',
   styleUrl: './shelf-management.component.scss'
 })
-//ngOnInit() ?
+
 export class ShelfManagementComponent implements OnInit {
 
-  //shelves : Shelf[] = [];
+  searchForm = this.fb.nonNullable.group({
+    searchText: [''] // Arama metni için değişken eklendi
+  });
+
+  itemsPerPage= 8;
   selectedShelf: AdminShelf | null = null;
-  // selectTableShelf: Shelf | null = null;
+  selvesPerPage: AdminShelf[] = [];
+  areYouSureQuestion = 'Are you sure you want to edit this shelf ?'
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
-     private shelfService: ShelfService,
-     private toastr: ToastrService,
-     private http: HttpClient
-  ) { }
+    private shelfService: ShelfService,
+    private toastr: ToastrService,
+    private fb: FormBuilder
+  ) {}
 
   shelves: AdminShelf[] = [];
   products: AdminProduct[] = [];
 
-  // ngOnInit(): void {
-  //   this.http.get<any[]>('/getAllShelves').subscribe(
-  //     data => {
-  //       console.log(data);
-  //       this.shelves = data;
-  //     },
-  //     error => {
-  //       console.error('Error fetching shelves:', error);
-  //     }
-  //   );
-  // }
-
   ngOnInit(): void {
+    this.loadShelves();
+    this.searchForm.get("searchText")?.valueChanges.subscribe({
+      next: (data) => {
+        this.filterShelf(data);
+      }
+    })
+  }
+
+  
+  loadShelves(){
     this.shelfService.getAllTableShelves().subscribe({
       next: (shelf => {
-        console.log(shelf);
         this.shelves = shelf;
+        this.totalPages= Math.ceil(this.shelves.length/this.itemsPerPage)
+        this.updatePageShelves();
+        this.filterShelf();
       })
     });
   }
 
-  //Component çağrıldığında çalışan method.
-  // ngOnInit(): void {
-  //   this.shelfService.getAllShelves().subscribe({
-  //     next: (data => {
-  //       this.shelves = data;
-  //   console.log(this.shelves);
-  //     })
-  //   });
-  // }
+  filteredShelves: AdminShelf[] = [];
+  filterShelf(data = "") {
+    this.filteredShelves = this.shelves;
+    // Arama filtresi
+    if (data) {
+      this.filteredShelves = this.filteredShelves.filter(shelf => {
+                return shelf.productCategory?.toLowerCase().includes(data.toLowerCase())
+      }
+      );
+    } else {
+      this.filteredShelves = this.shelves;
+    }
+    this.totalPages = Math.ceil(this.filteredShelves.length / this.itemsPerPage);
+    this.currentPage = 1; 
+    this.updatePageShelves();
+  }
   
   addShelf(){
     this.router.navigate(['addShelf'], { relativeTo: this.route });
   }
 
   setSelectedShelf(shelf: AdminShelf) {
-    // console.log(this.selectedShelf);
-    // if (shelf == this.selectedShelf) {
-    //   this.selectedShelf = null;
-    // } else if(shelf.productCount > 0){
-    //   this.selectedShelf = shelf;
-    //   this.getSelectedShelfProducts();
-    // } 
-
     if(shelf === this.selectedShelf){
       // Seçili raf zaten varsa ve tekrar tıklanırsa, seçili rafı kapat
       this.selectedShelf = null;
     } else {
       // Seçili rafı değiştir
       this.selectedShelf = shelf;
-      console.log(this.selectedShelf);
       this.getSelectedShelfProducts();
     }
   }
@@ -93,32 +94,69 @@ export class ShelfManagementComponent implements OnInit {
     }
   }
 
-  // selectShelf(shelf: Shelf) {
-  //   if (shelf == this.selectTableShelf) {
-  //     this.selectTableShelf = null;
-  //   } else {
-  //     this.selectTableShelf = shelf;
-  //   }
-  // }
-
   editShelf(shelf : AdminShelf) {
     this.shelfService.editingShelf = shelf;
-    console.log(shelf);
     this.router.navigate(['editShelf'], { relativeTo: this.route });
   }
   
-  deleteShelf(shelf: AdminShelf) {
-    console.log(shelf.id);
-    this.shelfService.deleteShelf(shelf.id).subscribe({
+  deleteShelf(shelf: AdminShelf | null) {
+    if(shelf){
+      this.shelfService.deleteShelf(shelf.id).subscribe({
       next: () => {
         this.shelves = this.shelves.filter(s => s.id!== shelf.id);
-        this.toastr.success("Shelf deleted successfully");
-        console.log(this.shelves);
+        this.filteredShelves = this.filteredShelves.filter(s => s.id !== shelf.id);
+        this.toastr.success("Shelf Successfully Deleted !");
       },
       error: (err)=> {
         console.log(err);
       }
-    })
+      })
+    }
+  }
+  updatePageShelves(): void {
+    const startIndex = (this.currentPage - 1) * this.itemsPerPage;
+    this.selvesPerPage = this.filteredShelves.slice(startIndex, startIndex + this.itemsPerPage);
+  }
+  
+
+  totalPages = 0;
+  currentPage = 1;
+  onPageChange(pageNo: number) {
+    if (pageNo < 1 || pageNo > this.totalPages) {
+      return; 
+    }
+    this.currentPage = pageNo;
+    this.updatePageShelves();
   }
 
+
+  totalPagesArray(): number[] {
+    const numPagesToShow = 5; 
+  const currentPage = this.currentPage;
+  const totalPages = this.totalPages;
+  
+  let startPage: number;
+  let endPage: number;
+
+  if (totalPages <= numPagesToShow) {
+    
+    startPage = 1;
+    endPage = totalPages;
+  } else {
+   
+    if (currentPage <= Math.floor(numPagesToShow / 2)) {
+   
+      startPage = 1;
+      endPage = numPagesToShow;
+    } else if (currentPage + Math.floor(numPagesToShow / 2) >= totalPages) {
+     
+      startPage = totalPages - numPagesToShow + 1;
+      endPage = totalPages;
+    } else {
+      startPage = currentPage - Math.floor(numPagesToShow / 2);
+      endPage = currentPage + Math.floor(numPagesToShow / 2);
+    }
+  }
+  return Array(endPage - startPage + 1).fill(0).map((_, index) => startPage + index);
+  }
 }
